fix(kitty): italics font
I guess it was because an inappropiate font has been used. I forgot what
`propo` does, but removing it seems to be fixing the italic font that
wasn't rendered. Perhaps the name was font name I put there was wrong or
something.

This commit also tweaks the colors in the visuals.